slide_analyzer/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.44.2
timm>=0.9.0
accelerate>=0.20.0

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0

# SAM (Segment Anything Model)
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# OpenAI API
openai>=1.0.0
